wip: split recommended apps surface
